requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.47.0
tiktoken==0.7.0
faiss-cpu==1.8.0
sqlite-utils==3.36